+++ lib/components/features/EventList.tsx
@@ -9,7 +9,6 @@ import { Address, Event, SpaceTagBase, User } from '$lib/generated/backend/graph
 import { generateUrl } from '$lib/utils/cnd';
 import { userAvatar } from '$lib/utils/user';
 import { getEventPrice } from '$lib/utils/event';
-import { ASSET_PREFIX } from '$lib/utils/constants';
 
 export function EventList({
   events,
@@ -317,7 +316,7 @@ function getLocation(address?: Address | null) {
 function EmptyComp() {
   return (
     <div className="flex flex-col items-center flex-1 gap-5 mt-10 mb-20">
-      <img src={`${ASSET_PREFIX}/assets/images/ic_dashboard.png`} />
+      <i className="icon-dashboard w-[184px] h-[184px] text-primary/16 blur-md" />
       <div className="text-center">
         <p className="font-title text-tertiary font-semibold! text-xl">No Upcoming Events</p>
         <p className="text-tertiary">Subscribe to keep up with new events.</p>